integer amount

+++ src/main/java/sk/stuba/fei/uim/oop/assignment3/product/Product.java
@@ -14,7 +14,7 @@ public class Product {
     private Long id;
     private String name;
     private String description;
-    private Long amount;
+    private int amount;
     private String unit;
     private Double price;
 }
